ai_pipeline/requirements.txt
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=11.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
pandas>=2.0.0
tqdm>=4.65.0
sentencepiece>=0.1.99
